prime-trade-ai/requirements.txt
python-binance==1.0.19
colorama==0.4.6
python-dotenv==1.0.0
